feat(management): Replace TomorrowMorning with FollowingMorning
'Tomorrow' is defined in the code as today (date of now), plus one
day. This can lead to confusion as some people would count tomorrow
only from the time they wake up from sleep / only from morning.

Thus Tomorrow morning is changed to following morning to avoid this
confusion, and the logic is changed to find the next morning, ie.
either morning today if it is between midnight and 6, or
tomorrow 6 if it is past 6.