tests(ctuauth): add KosApi, UsermapApi source builders
This is for populating the apis with data, along with dummy api
implementations that will be used during testing.

There is a simple builder that can be used for creating faculties,
people, their student and teacher roles.

This should already have everything needed for testing the
authentication flow, other types could be added, but they
are not necessary.

The output of the `CtuSourceBuilder` are two sources:
- for usermap, a  dictionary of usernames and UsermapPerson as values
- for kosapi, a dictionary  of faculties, programmes, people, students and teachers

The dummy apis are then taking those sources.

For the kosapi, only the atom api is made. This is sufficient as
implementations of other classes all call atom api.

For the usermap api, only person api is made, as the other api
isn't used.